refactor: rename project to parallel-corpus
