+++ .github/workflows/release.yml
@@ -88,10 +88,10 @@ jobs:
           name: pypi_files
           path: dist
 
-      - run: rm -r <INSERT PROJECT SRC>
+      - run: rm -r src/parallel_corpus
       - run: pip install typing-extensions
       - run: pip install -r tests/requirements-testing.lock
-      - run: pip install <INSERT PROJECT NAME> --no-index --no-deps --find-links dist --force-reinstall
+      - run: pip install parallel-corpus --no-index --no-deps --find-links dist --force-reinstall
       - run: pytest
   
   # https://github.com/marketplace/actions/alls-green#why used for branch protection checks
