+++ src/parallel_corpus/shared/functional.py
