+++ tests/test_token.py
@@ -1,7 +1,7 @@
 from typing import List
 
 import pytest
-from graph.token import Token, identify, tokenize
+from parallel_corpus.token import Token, identify, tokenize
 
 
 def test_can_create_token() -> None:
